--- back-end/app/api/src/instrumentation.ts
@@ -1,6 +1,0 @@
-// import connectDB from "../database/database";
-import connectMongo from "../database/database";
-
-export async function register() {
-  await connectMongo();
-}
