+++ back-end/src/app/api/product/route.ts
@@ -1,40 +1,44 @@
 import { HttpStatusCode } from "axios";
-import UserModel from "../../models/user.model";
+
+import ProductModel from "../../../models/product.model";
 import { NextRequest, NextResponse } from "next/server";
+
+type CreateProductDto = {
+  name: string;
+  image: string;
+  color: string;
+  brand: string;
+  gender: string;
+  price: number;
+  discount: string;
+  rest: number;
+  size: number;
+};
 
 export async function POST(req: NextRequest) {
   try {
-    type CreateProductDto = {
-      name: String;
-      password: Number;
-      firstName: String;
-      lastName: String;
-      phoneNumber: Number;
-      role: String;
-      orderId: String[];
-    };
+    console.log("product deer huselt irlee");
     const body: CreateProductDto = await req.json();
-    const { name, password, firstName, lastName, phoneNumber, role, orderId } =
+    const { name, image, color, brand, gender, price, discount, rest, size } =
       body;
+
     if (
       name &&
-      password &&
-      firstName &&
-      lastName &&
-      phoneNumber &&
-      role &&
-      orderId
+      image &&
+      price &&
+      color &&
+      brand &&
+      gender &&
+      discount &&
+      rest &&
+      size
     ) {
-      const user = await UserModel.create(body);
+      const product = await ProductModel.create(body);
       return NextResponse.json(
-        { user, message: "Your user has been created" },
+        { product, message: "Your product has been created" },
         { status: HttpStatusCode.Created }
       );
     }
-    return NextResponse.json(
-      { message: "User name is missing" },
-      { status: HttpStatusCode.BadRequest }
-    );
   } catch (error) {
     console.log(error);
     return NextResponse.json(
@@ -45,7 +49,7 @@ export async function POST(req: NextRequest) {
 }
 export async function GET(_: NextRequest) {
   try {
-    const product = await UserModel.find();
+    const product = await ProductModel.find();
     if (product) {
       return NextResponse.json({ product });
     }
