+++ back-end/src/app/favicon.ico
