+++ back-end/src/instrumentation.ts
@@ -1,0 +1,5 @@
+import connectMongo from "./database/database";
+
+export async function register() {
+  await connectMongo();
+}
