+++ back-end/src/models/order.model.ts
@@ -5,7 +5,6 @@ const OrderSchema = new Schema(
     userId: { type: [Schema.Types.ObjectId], ref: COLLECTIONS.USER },
     orderNumber: { type: Number, required: true },
     product: { type: [Schema.Types.ObjectId], ref: COLLECTIONS.PRODUCT },
-    totalPrice: { type: Number, required: true },
     status: {
       type: String,
       enum: ["Hurgegdsen", "Hurgegdeegui", "HurgeltendGarsan"],
